Remove `register.py` files in favor of finer granularity
