+++ src/python/pants/engine/exp/examples/BUILD
@@ -25,7 +25,6 @@ python_library(
     'src/python/pants/engine/exp:graph',
     'src/python/pants/engine/exp:mapper',
     'src/python/pants/engine/exp:nodes',
-    'src/python/pants/engine/exp:register',
     'src/python/pants/engine/exp:scheduler',
     'src/python/pants/engine/exp:selectors',
     'src/python/pants/engine/exp:storage',
